# Rewrite MCU I2C to support non blocking execution.