--- a/src/avr/gpio.h
+++ b/src/avr/gpio.h
@@ -2,6 +2,7 @@
 #define __AVR_GPIO_H
 
 #include <stdint.h>
+#include "sched.h" // timer
 
 struct gpio_out {
     struct gpio_digital_regs *regs;
@@ -53,5 +54,6 @@ struct i2c_config i2c_setup(uint32_t bus, uint32_t rate, uint8_t addr);
 void i2c_write(struct i2c_config config, uint8_t write_len, uint8_t *write);
 void i2c_read(struct i2c_config config, uint8_t reg_len, uint8_t *reg
               , uint8_t read_len, uint8_t *read);
+uint_fast8_t i2c_async(struct timer *timer);
 
 #endif // gpio.h