--- a/src/avr/gpio.h
+++ b/src/avr/gpio.h
@@ -48,6 +48,7 @@ void spi_transfer(struct spi_config config, uint8_t receive_data
 
 struct i2c_config {
     uint8_t addr;
+    uint32_t pause;
 };
 
 struct i2c_config i2c_setup(uint32_t bus, uint32_t rate, uint8_t addr);