--- a/src/i2c_software.c
+++ b/src/i2c_software.c
@@ -179,3 +179,35 @@ i2c_software_read(struct i2c_software *is, uint8_t reg_len, uint8_t *reg
     }
     i2c_software_stop(is);
 }
+
+uint_fast8_t i2c_software_async(struct timer *timer)
+{
+    struct i2cdev_s *i2c = container_of(timer, struct i2cdev_s, timer);
+    struct i2c_software *is = alloc_chunk(sizeof(*is));
+    if (i2c->data_len[1] & I2C_R) {
+        uint8_t reg_len = i2c->data_len[0];
+        uint8_t *reg = i2c->buf;
+        uint8_t read_len = i2c->data_len[1] & ~(I2C_R);
+        uint8_t *resp = &i2c->buf[reg_len];
+        i2c_software_read(is, reg_len, reg, read_len, resp);
+
+        i2c->cur = i2c->tail + read_len;
+        timer->func = i2c->callback;
+        timer->waketime = timer_read_time() + timer_from_us(50);
+        return SF_RESCHEDULE;
+    } else if (i2c->data_len[0]) {
+        uint8_t to_write = i2c->data_len[0];
+        uint8_t buf[to_write];
+        for (int i = 0; i < to_write; i++)
+        {
+            buf[i] = i2c_buf_read_b(i2c);
+        }
+
+        i2c_software_write(is, to_write, buf);
+        timer->waketime = timer_read_time() + timer_from_us(50);
+        return SF_RESCHEDULE;
+    }
+
+    i2c->flags &= ~IF_ACTIVE;
+    return SF_DONE;
+}