--- a/src/linux/gpio.h
+++ b/src/linux/gpio.h
@@ -47,6 +47,7 @@ void gpio_pwm_write(struct gpio_pwm g, uint16_t val);
 struct i2c_config {
     int fd;
     uint8_t addr;
+    int ret;
 };
 
 struct i2c_config i2c_setup(uint32_t bus, uint32_t rate, uint8_t addr);