… src/lpc176x/gpio.h
… src/lpc176x/i2c.c
… src/rp2040/gpio.h
… src/rp2040/i2c.c
… src/stm32/gpio.h
… src/stm32/i2c.c
… src/stm32/stm32f0_i2c.c