@KevinOConnor 2024-09-10
Interesting, thanks.

There seems to be a lot of work going on here.  Out of curiosity, what is the underlying issue that is being addressed?  For example, is there some hardware that needs high-speed access that the current code can't support?

Some random thoughts:
1. The reason for the high "mcu load" graphs is due to variability in the time it takes to idle when i2c devices are used.  That is, the "mcu load" statistic is based on the total time the mcu spends running tasks (` mcu_awake`) and the variability in the timing to complete all pending tasks (`mcu_task_stddev`).  When the host sends periodic i2c requests it results in high variability, as most of the time tasks complete in a few microseconds, but when an i2c command is pending it may take a millisecond or so.  Variability is frequently a symptom of high load (which is why it is used in the load calculation), but it isn't inherently a problem.  Tasks are low-priority and as long as the delays are under a millisecond or so, there shouldn't be a degradation in performance.
2. The `i2c_modify_bits` command is only used in sx1509.py, and that code doesn't require it.  Likely better to remove `i2c_modify_bits` than to spend lots of time reworking it.
3. If you change the `i2c_read` command to be asynchronous then the host code needs to also change.  The host code expects the results of `i2c_read` to always be received synchronously with its command ack.  (Note the `is_async` flag in `lookup_query_command`.)

Cheers,
-Kevin