@nefelim4ag 2024-09-10
My main motivation here was precious step timings _and a little dislike for `while()` loops_.
The only serious reason here, is that all TMC features use timings and data between steps, and when scheduler lags behind "now" and trying to catch up - that can confuse driver functions.

TMC drivers don't have(2209) or have small hysteresis for TSTEP, instantaneous change of TSTEP can cause lags of interpolation, instantaneous enabling, and disabling of stealth chop & other speed-related functionality.
_But this is actually hard to prove._
That's it, no user-visible issue here, other stuff is actually not so important I think (like graphs).

_(Maybe this is the wrong approach to such a theoretical problem, and just computing and limiting minimal time between steps, based on machine speed limits can be enough...)._

I2C is just slightly faster than USART, but still can take a measurable amount of time.

My changes does not speedup i2c, so i2c devices will behave as before in terms of speed.
That changes only tradeoff code memory/complexity for CPU free time & tail latency.

There are also SPI coprocessors, but as mentioned above, crazy fast, even if we will poll driver status every ms, it is unlikely it will affect anything.

1. Understood.
2. Agree. I just can't suggest that without your approval and I did not have such expander to test. But I can open PR with such changes to sx1509, np.
3. You are right, but I already set is_async=True for i2c_read, to avoid issues with resending before responding and broken communication orders, this one just increased timeouts. Did I miss something else?

Thanks for your time.

This is completely fine if you say this is too much overcomplication, and I can just extract useful bits from here:
- Correct end of SW I2C transfer on NACK.
- SX1509 modification and depreciation of i2c_modify_bits().