@KevinOConnor 2024-09-11
Thanks.

I don't object to you (or others) reworking the i2c implementation.  From my perspective, the Klipper i2c implementation is a "hack" - I added it to support some DAC chips that configured stepper current on old boards - and it has since grown to a number of unexpected configurations.

FWIW, one significant limitation of the current i2c implementation is that it calls `shutdown()` on minor errors.  In retrospect, it would have been preferable for both `i2c_read()` and `i2c_write()` to always respond to requests with a status (success or failure) and then let the host figure out how to handle the problem.

> My main motivation here was precious step timings

I'm confused by this statement.  Stepper step pulses are generated from the irq handlers which are high priority events.  They will preempt Klipper "tasks" (such as the i2c command handler) and thus i2c should not impact step timing.

I'm also not clear on the relationship between i2c and `TSTEP`.

> You are right, but I already set is_async=True for i2c_read

Oh, sorry - I missed that in your changes.

Cheers,
-Kevin