@nefelim4ag 2024-09-11
Nice hack BTW :D

> significant limitation of the current i2c implementation is that it calls shutdown() on minor errors.

Yes, I2C errors right now are a little annoying sometimes, in comparison to SPI/ADXL which just returns 0x00/0xff.
I will be happy to fix that.

_BTW, STM32 missing check for I2C_ISR_NACKF_

Looking at SDIO,
technically we can do the same - define "error=%c" or "status=%c" field, define enum with "typical" errors.
Then just do raise i2c_error_X() (in bus.py) and allow handling of it in high level code.
I can do that in other PR.

Least we can do actually, in case of errors, is only retry on NACK.
high level stuff like reinitialization of device seems, strange.

I'm a little afraid that this will lead to a little storm of patches with little fixes for bad wiring.

> I'm confused by this statement.

That is my misunderstanding then, I somewhat thought that timers run periodic tasks, and as a consequence, they also have the same priority. So command.c -> i2c_(read/write) & other stuff are effectively blocking everything else.

So, as steps are unaffected, now this more like competitive programming and now I'm not sure that it is worth it :\
Because it will only make nice graphs, and free a very little CPU time.
is completely up to your, does it still has any value.

Thanks.