@KevinOConnor 2024-09-13
> I somewhat thought that timers run periodic tasks, and as a consequence, they also have the same priority

Okay, that's not the case.  Most of the Klipper mcu code is contained in timers or tasks, and timers run at a higher priority than tasks.  It's okay to have lot of tasks and long running tasks because they do not cause increased latency for timers.

> is completely up to your, does it still has any value.

As before, it's fine with me if someone wants to improve the i2c handling.  It's not a performance issue (as far as I know), but I'm fine with incremental changes that make improvements.

Cheers,
-Kevin