@nefelim4ag 2025-02-08
For the history, this PR do a minor improvements, but there is no practical reason to do so with this level of added complexity.